requirements.txt
torch
torchvision
pandas
Pillow
flask
